requirements.txt
fastapi[standard]
pydantic
uvicorn
gtts
openai-whisper
torch
numpy